Update amount in cents format. Add initial migration to create the database and services table.

+++ Models/Domains/Service.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal MinPrice { get; set; }
-        public decimal MaxPrice { get; set; }
+        public int MinPrice { get; set; }
+        public int MaxPrice { get; set; }
         public int Status { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
